Move converters to own namespace
