+++ BuildNotifications/Resources/Global/Navigation/ButtonNavigation/ButtonNavigationTemplate.xaml
@@ -8,6 +8,7 @@
                     xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
                     xmlns:utils="clr-namespace:BuildNotifications.ViewModel.Utils"
                     xmlns:system="clr-namespace:System;assembly=System.Runtime"
+                    xmlns:converters="clr-namespace:BuildNotifications.ViewModel.Utils.Converters"
                     mc:Ignorable="d">
     <Style TargetType="{x:Type buttonNavigation:ButtonNavigation}">
         <Setter Property="Template">
@@ -18,7 +19,7 @@
 	                        <icons:IconButton Icon="Plus" Command="{TemplateBinding AddItemCommand}" x:Name="IconBtn"
 	                                          Foreground="{DynamicResource Background3}" DockPanel.Dock="Bottom"
 	                                          Background="{DynamicResource Background1}"
-	                                          Visibility="{TemplateBinding AddItemCommand, Converter={x:Static utils:NullToVisibilityConverter.Instance}}"
+	                                          Visibility="{TemplateBinding AddItemCommand, Converter={x:Static converters:NullToVisibilityConverter.Instance}}"
 	                                          Height="{StaticResource Block}" IconSize="{StaticResource Block}">
 		                        <i:Interaction.Triggers>
 			                        <i:EventTrigger EventName="Click">
@@ -40,14 +41,14 @@
                                                              Foreground="{DynamicResource Foreground1HalfTransparency}"
                                                              Width="{StaticResource BlockOneAndHalf}"
                                                              Height="{StaticResource BlockOneAndHalf}"
-                                                             Visibility="{Binding Icon, Converter={x:Static utils:IconTypeToVisibilityConverter.Instance}}"
+                                                             Visibility="{Binding Icon, Converter={x:Static converters:IconTypeToVisibilityConverter.Instance}}"
                                                              />
 
                                             <icons:IconButton Icon="Minus" Margin="{StaticResource BlockHalfXOnlyMargin}"
                                                               Command="{Binding RelativeSource={RelativeSource AncestorType={x:Type buttonNavigation:ButtonNavigation}}, Path=RemoveItemCommand}"
                                                               CommandParameter="{Binding}" x:Name="IconBtn"
                                                               Foreground="{DynamicResource Background3}" DockPanel.Dock="Right"
-                                                              Visibility="{Binding RelativeSource={RelativeSource AncestorType={x:Type buttonNavigation:ButtonNavigation}}, Path=RemoveItemCommand, Converter={x:Static utils:NullToVisibilityConverter.Instance}}"
+                                                              Visibility="{Binding RelativeSource={RelativeSource AncestorType={x:Type buttonNavigation:ButtonNavigation}}, Path=RemoveItemCommand, Converter={x:Static converters:NullToVisibilityConverter.Instance}}"
                                                               Height="{StaticResource Block}" IconSize="{StaticResource Block}">
 	                                            <i:Interaction.Triggers>
 		                                            <i:EventTrigger EventName="Click">
