+++ BuildNotifications/ViewModel/Utils/Converters/BoolToVisibilityConverter.cs
@@ -3,7 +3,7 @@ using System.Globalization;
 using System.Windows;
 using System.Windows.Data;
 
-namespace BuildNotifications.ViewModel.Utils
+namespace BuildNotifications.ViewModel.Utils.Converters
 {
     internal class BoolToVisibilityConverter : IValueConverter
     {
