+++ BuildNotifications/Views/BusyIndicator.xaml
@@ -10,6 +10,7 @@
              xmlns:animation="clr-namespace:BuildNotifications.Resources.Animation"
              xmlns:text="clr-namespace:BuildNotifications.Resources.Text"
              xmlns:utils="clr-namespace:BuildNotifications.ViewModel.Utils"
+             xmlns:converters="clr-namespace:BuildNotifications.ViewModel.Utils.Converters"
              mc:Ignorable="d"
              d:DesignHeight="450" d:DesignWidth="800">
     <Grid Background="{DynamicResource Background1}" Margin="{StaticResource BlockXOnlyMargin}">
@@ -19,7 +20,7 @@
                       ItemsSource="{Binding DummyItems, RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type local:BusyIndicator}}}">
             <ItemsControl.ItemTemplate>
                 <DataTemplate>
-                    <Grid  Visibility="{Binding ShowLoadingText, Converter={x:Static utils:BoolToVisibilityConverter.Inverted}, RelativeSource={RelativeSource AncestorType={x:Type local:BusyIndicator}}}">
+                    <Grid  Visibility="{Binding ShowLoadingText, Converter={x:Static converters:BoolToVisibilityConverter.Inverted}, RelativeSource={RelativeSource AncestorType={x:Type local:BusyIndicator}}}">
                         <Rectangle Width="{DynamicResource BlockTriple}" Height="{DynamicResource Block}"
                                    Margin="{DynamicResource BlockHalfThickness}"
                                    Fill="{DynamicResource Foreground2}">
@@ -64,7 +65,7 @@
             </i:Interaction.Triggers>
         </ItemsControl>
 
-        <TextBlock Text="{text:Localized Loading}" Foreground="{DynamicResource Foreground1}" HorizontalAlignment="Right" Visibility="{Binding ShowLoadingText, Converter={x:Static utils:BoolToVisibilityConverter.Instance}, RelativeSource={RelativeSource AncestorType={x:Type local:BusyIndicator}}}"/>
+        <TextBlock Text="{text:Localized Loading}" Foreground="{DynamicResource Foreground1}" HorizontalAlignment="Right" Visibility="{Binding ShowLoadingText, Converter={x:Static converters:BoolToVisibilityConverter.Instance}, RelativeSource={RelativeSource AncestorType={x:Type local:BusyIndicator}}}"/>
         <i:Interaction.Triggers>
             <ei:DataTrigger
                 Binding="{Binding Path=IsBusy, RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type local:BusyIndicator}}}"
