+++ BuildNotifications/Views/InfoPopupDialog.xaml
@@ -6,7 +6,7 @@
                      xmlns:window="clr-namespace:BuildNotifications.Resources.Window"
                      xmlns:viewModel="clr-namespace:BuildNotifications.ViewModel"
                      xmlns:text="clr-namespace:BuildNotifications.Resources.Text"
-                     xmlns:utils="clr-namespace:BuildNotifications.ViewModel.Utils"
+                     xmlns:converters="clr-namespace:BuildNotifications.ViewModel.Utils.Converters"
                      mc:Ignorable="d" WindowStartupLocation="CenterOwner" Width="300" Height="300"
                      d:DataContext="{d:DesignInstance d:Type={x:Type viewModel:InfoPopupViewModel}}">
 
@@ -35,19 +35,19 @@
         <TextBlock />
 
         <text:ClickableTextBlock Text="{text:Localized CheckForUpdates}" HorizontalAlignment="Center"
-                                 Visibility="{Binding CheckForUpdatesCommand.Execution, Converter={x:Static utils:NullToVisibilityConverter.Inverted}}"
+                                 Visibility="{Binding CheckForUpdatesCommand.Execution, Converter={x:Static converters:NullToVisibilityConverter.Inverted}}"
                                  Command="{Binding CheckForUpdatesCommand}" />
         <Grid
-            Visibility="{Binding CheckForUpdatesCommand.Execution, Converter={x:Static utils:NullToVisibilityConverter.Instance}}">
+            Visibility="{Binding CheckForUpdatesCommand.Execution, Converter={x:Static converters:NullToVisibilityConverter.Instance}}">
             <TextBlock Text="{text:Localized CheckingForUpdates}" HorizontalAlignment="Center"
-                       Visibility="{Binding CheckForUpdatesCommand.Execution.IsNotCompleted, Converter={x:Static utils:BoolToVisibilityConverter.Instance}}" />
+                       Visibility="{Binding CheckForUpdatesCommand.Execution.IsNotCompleted, Converter={x:Static converters:BoolToVisibilityConverter.Instance}}" />
 
-            <Grid Visibility="{Binding CheckForUpdatesCommand.Execution.IsSuccessfullyCompleted, Converter={x:Static utils:BoolToVisibilityConverter.Instance}}">
+            <Grid Visibility="{Binding CheckForUpdatesCommand.Execution.IsSuccessfullyCompleted, Converter={x:Static converters:BoolToVisibilityConverter.Instance}}">
                 <TextBlock Text="{text:Localized NoUpdatesAvailable}" HorizontalAlignment="Center"
-                           Visibility="{Binding CheckForUpdatesCommand.Execution.Result, Converter={x:Static utils:BoolToVisibilityConverter.Inverted}}" />
+                           Visibility="{Binding CheckForUpdatesCommand.Execution.Result, Converter={x:Static converters:BoolToVisibilityConverter.Inverted}}" />
 
                 <text:ClickableTextBlock Text="{text:Localized UpdateAvailable}" Command="{Binding UpdateCommand}" HorizontalAlignment="Center"
-                                         Visibility="{Binding CheckForUpdatesCommand.Execution.Result, Converter={x:Static utils:BoolToVisibilityConverter.Instance}}" />
+                                         Visibility="{Binding CheckForUpdatesCommand.Execution.Result, Converter={x:Static converters:BoolToVisibilityConverter.Instance}}" />
             </Grid>
         </Grid>
