+++ BuildNotifications/Views/Tree/BuildTreeView.xaml
@@ -10,6 +10,7 @@
              xmlns:utils="clr-namespace:BuildNotifications.ViewModel.Utils"
              xmlns:icons="clr-namespace:BuildNotifications.Resources.Icons"
              xmlns:text="clr-namespace:BuildNotifications.Resources.Text"
+             xmlns:converters="clr-namespace:BuildNotifications.ViewModel.Utils.Converters"
              mc:Ignorable="d"
              d:DesignHeight="450" d:DesignWidth="800"
              d:DataContext="{d:DesignInstance Type={x:Type viewModelsTree:BuildTreeViewModel}}">
@@ -35,7 +36,7 @@
             </ItemsControl>
         </ScrollViewer>
 
-        <DockPanel Visibility="{Binding TreeIsEmpty, Converter={x:Static utils:BoolToVisibilityConverter.Instance}}"
+        <DockPanel Visibility="{Binding TreeIsEmpty, Converter={x:Static converters:BoolToVisibilityConverter.Instance}}"
                    HorizontalAlignment="Center" VerticalAlignment="Center">
             <icons:IconBatch Type="NoConnection" Width="{StaticResource BlockTriple}"
                              Height="{StaticResource BlockTriple}" Foreground="{StaticResource Background3}"
